Psl 194 convert pastel gtest (#97)
* convert test_accounting.cpp

* clean

* fix test_accounting.cpp issue

* finish test_alert.cpp

* convert bloom_tests.cpp

* remove unused target Makeilfe

* fix findings

* remvoe comment out

* remove unused var

* improve method

* fix circelcl failed on raw generated data
