+++ src/Makefile.gtest.include
@@ -1,6 +1,10 @@
 TESTS += pastel-gtest
 noinst_PROGRAMS += pastel-gtest
 
+RAW_GTEST_FILES = gtest/data/alertTests.raw
+
+GENERATED_GTEST_FILES = $(RAW_GTEST_FILES:.raw=.raw.h)
+
 # tool for generating our public parameters.
 # test_checktransaction.cpp MUST be before
 # any test that calls SelectParams().
@@ -30,6 +34,7 @@ pastel_gtest_SOURCES +=\
 	gtest/test_mnode/test_ticket_mempool_processor.cpp\
 	gtest/test_mnode/test_ticket_processor.cpp\
 	gtest/test_addrman.cpp\
+	gtest/test_alert.cpp\
 	gtest/test_allocator.cpp\
 	gtest/test_arith_uint256.cpp\
 	gtest/test_base32.cpp\
@@ -38,6 +43,7 @@ pastel_gtest_SOURCES +=\
 	gtest/test_bech32.cpp\
 	gtest/test_bip32.cpp\
 	gtest/test_block.cpp\
+	gtest/test_bloom.cpp\
 	gtest/test_checkblock.cpp\
 	gtest/test_checkpoints.cpp\
 	gtest/test_circuit.cpp\
@@ -89,12 +95,12 @@ pastel_gtest_SOURCES +=\
 	
 if ENABLE_WALLET
 pastel_gtest_SOURCES += \
+	wallet/gtest/test_accounting.cpp\
 	wallet/gtest/test_wallet.cpp\
 	wallet/gtest/test_wallet_ismine.cpp
 endif
 
-
-pastel_gtest_CPPFLAGS = $(AM_CPPFLAGS) -DBINARY_OUTPUT -DCURVE_ALT_BN128 -DSTATIC $(BITCOIN_INCLUDES) -I$(srcdir)/gtest
+pastel_gtest_CPPFLAGS = $(AM_CPPFLAGS) -DBINARY_OUTPUT -DCURVE_ALT_BN128 -DSTATIC $(BITCOIN_INCLUDES) -I$(srcdir)/gtest 
 pastel_gtest_CXXFLAGS = $(AM_CXXFLAGS) $(PIE_FLAGS) -I$(srcdir)/gtest
 
 pastel_gtest_LDADD =\
@@ -136,8 +142,20 @@ endif
 
 pastel_gtest_LDFLAGS = $(RELDFLAGS) $(AM_LDFLAGS) $(LIBTOOL_APP_LDFLAGS) -static
 
+nodist_pastel_gtest_SOURCES = $(GENERATED_GTEST_FILES)
+
+$(pastel_gtest_SOURCES): $(GENERATED_GTEST_FILES)
+
 pastel-gtest_check: pastel-gtest FORCE
 	./pastel-gtest
 
 pastel-gtest-expected-failures: pastel-gtest FORCE
 	./pastel-gtest --gtest_filter=*DISABLED_* --gtest_also_run_disabled_tests
+
+%.raw.h: %.raw
+	@$(MKDIR_P) $(@D)
+	@echo "namespace alert_tests{" > $@
+	@echo "static unsigned const char $(*F)[] = {" >> $@
+	@$(HEXDUMP) -v -e '8/1 "0x%02x, "' -e '"\n"' $< | $(SED) -e 's/0x  ,//g' >> $@
+	@echo "};};" >> $@
+	@echo "Generated $@"
