+++ src/Makefile.test.include
@@ -39,14 +39,10 @@ JSON_TEST_FILES = \
   test/data/g2_compressed.json \
   test/data/sapling_key_components.json
 
-RAW_TEST_FILES = test/data/alertTests.raw
-
-GENERATED_TEST_FILES = $(JSON_TEST_FILES:.json=.json.h) $(RAW_TEST_FILES:.raw=.raw.h)
+GENERATED_TEST_FILES = $(JSON_TEST_FILES:.json=.json.h)
 
 BITCOIN_TESTS =\
   test/bignum.h \
-  test/alert_tests.cpp \
-  test/bloom_tests.cpp \
   test/checkblock_tests.cpp \
   test/coins_tests.cpp \
   test/dbwrapper_tests.cpp \
@@ -72,12 +68,11 @@ BITCOIN_TESTS =\
 
 if ENABLE_WALLET
 BITCOIN_TESTS += \
-  test/accounting_tests.cpp \
   wallet/test/wallet_tests.cpp \
   test/rpc_wallet_tests.cpp
 endif
 
-test_test_bitcoin_SOURCES = $(BITCOIN_TESTS) $(JSON_TEST_FILES) $(RAW_TEST_FILES)
+test_test_bitcoin_SOURCES = $(BITCOIN_TESTS) $(JSON_TEST_FILES)
 test_test_bitcoin_CPPFLAGS = $(AM_CPPFLAGS) $(BITCOIN_INCLUDES) -I$(builddir)/test/ $(TESTDEFS) $(EVENT_CFLAGS)
 test_test_bitcoin_LDADD =
 if ENABLE_WALLET
@@ -149,10 +144,3 @@ check-local:
 	@echo "};};" >> $@
 	@echo "Generated $@"
 
-%.raw.h: %.raw
-	@$(MKDIR_P) $(@D)
-	@echo "namespace alert_tests{" > $@
-	@echo "static unsigned const char $(*F)[] = {" >> $@
-	@$(HEXDUMP) -v -e '8/1 "0x%02x, "' -e '"\n"' $< | $(SED) -e 's/0x  ,//g' >> $@
-	@echo "};};" >> $@
-	@echo "Generated $@"
